+++ src/icloudpd/download.py
@@ -4,21 +4,20 @@ import os
 import socket
 import time
 import logging
+import datetime
 from tzlocal import get_localzone
 from requests.exceptions import ConnectionError  # pylint: disable=redefined-builtin
 from pyicloud_ipd.exceptions import PyiCloudAPIResponseError
-from icloudpd.logger import setup_logger
 
 # Import the constants object so that we can mock WAIT_SECONDS in tests
 from icloudpd import constants
 
-
-def update_mtime(photo, download_path):
+def update_mtime(created: datetime.datetime, download_path):
     """Set the modification time of the downloaded file to the photo creation date"""
-    if photo.created:
+    if created:
         created_date = None
         try:
-            created_date = photo.created.astimezone(
+            created_date = created.astimezone(
                 get_localzone())
         except (ValueError, OSError):
             # We already show the timezone conversion error in base.py,
@@ -33,33 +32,73 @@ def set_utime(download_path, created_date):
     ctime = time.mktime(created_date.timetuple())
     os.utime(download_path, (ctime, ctime))
 
+def mkdirs_for_path(logger, download_path: str) -> bool:
+    """ Creates hierarchy of folders for file path if it needed """
+    try:
+        # get back the directory for the file to be downloaded and create it if
+        # not there already
+        download_dir = os.path.dirname(download_path)
+        os.makedirs(name = download_dir, exist_ok=True)
+        return True
+    except OSError:
+        logger.tqdm_write(
+            f"Could not create folder {download_dir}",
+            logging.ERROR,
+        )
+        return False
 
-def download_media(icloud, photo, download_path, size):
-    """Download the photo to path, with retries and error handling"""
-    logger = setup_logger()
-
-    # get back the directory for the file to be downloaded and create it if
-    # not there already
+def mkdirs_for_path_dry_run(logger, download_path: str) -> bool:
+    """ DRY Run for Creating hierarchy of folders for file path """
     download_dir = os.path.dirname(download_path)
-
     if not os.path.exists(download_dir):
-        try:
-            os.makedirs(download_dir)
-        except OSError:  # pragma: no cover
-            pass         # pragma: no cover
+        logger.tqdm_write(
+            f"[DRY RUN] Would create folder hierarchy {download_dir}",
+            logging.DEBUG,
+        )
+    return True
+
+def download_response_to_path(
+        _logger,
+        response,
+        download_path: str,
+        created_date: datetime.datetime) -> bool:
+    """ Saves response content into file with desired created date """
+    temp_download_path = download_path + ".part"
+    with open(temp_download_path, "wb") as file_obj:
+        for chunk in response.iter_content(chunk_size=1024):
+            if chunk:
+                file_obj.write(chunk)
+    os.rename(temp_download_path, download_path)
+    update_mtime(created_date, download_path)
+    return True
+
+def download_response_to_path_dry_run(
+        logger,
+        _response,
+        download_path: str,
+        _created_date: datetime.datetime) -> bool:
+    """ Pretends to save response content into a file with desired created date """
+    logger.tqdm_write(
+        f"[DRY RUN] Would download {download_path}",
+        logging.INFO,
+    )
+    return True
+
+# pylint: disable-msg=too-many-arguments
+def download_media(logger, dry_run, icloud, photo, download_path, size) -> bool:
+    """Download the photo to path, with retries and error handling"""
+
+    mkdirs_local = mkdirs_for_path_dry_run if dry_run else mkdirs_for_path
+    download_local = download_response_to_path_dry_run if dry_run else download_response_to_path
+
+    if not mkdirs_local(logger, download_path):
+        return False
 
     for retries in range(constants.MAX_RETRIES):
         try:
             photo_response = photo.download(size)
             if photo_response:
-                temp_download_path = download_path + ".part"
-                with open(temp_download_path, "wb") as file_obj:
-                    for chunk in photo_response.iter_content(chunk_size=1024):
-                        if chunk:
-                            file_obj.write(chunk)
-                os.rename(temp_download_path, download_path)
-                update_mtime(photo, download_path)
-                return True
+                return download_local(logger, photo_response, download_path, photo.created)
 
             logger.tqdm_write(
                 f"Could not find URL to download {photo.filename} for size {size}!",
@@ -89,16 +128,18 @@ def download_media(icloud, photo, download_path, size):
                 time.sleep(wait_time)
 
         except IOError:
-            logger.error(
-                "IOError while writing file to %s! "
-                "You might have run out of disk space, or the file "
-                "might be too large for your OS. "
-                "Skipping this file...", download_path
+            logger.tqdm_write(
+                f"IOError while writing file to {download_path}! " +
+                "You might have run out of disk space, or the file " +
+                "might be too large for your OS. " +
+                "Skipping this file...", 
+                logging.ERROR
             )
             break
     else:
         logger.tqdm_write(
-            f"Could not download {photo.filename}! Please try again later."
+            f"Could not download {photo.filename}! Please try again later.", 
+            logging.ERROR,
         )
 
     return False