+++ src/icloudpd/exif_datetime.py
@@ -1,22 +1,21 @@
 """Get/set EXIF dates from photos"""
 
+import logging
 import piexif
 from piexif._exceptions import InvalidImageDataError
-from icloudpd.logger import setup_logger
 
 
-def get_photo_exif(path):
+def get_photo_exif(logger, path):
     """Get EXIF date for a photo, return nothing if there is an error"""
     try:
         exif_dict = piexif.load(path)
         return exif_dict.get("Exif").get(36867)
     except (ValueError, InvalidImageDataError):
-        logger = setup_logger()
-        logger.debug("Error fetching EXIF data for %s", path)
+        logger.tqdm_write(f"Error fetching EXIF data for {path}", logging.DEBUG)
         return None
 
 
-def set_photo_exif(path, date):
+def set_photo_exif(logger, path, date):
     """Set EXIF date on a photo, do nothing if there is an error"""
     try:
         exif_dict = piexif.load(path)
@@ -26,6 +25,5 @@ def set_photo_exif(path, date):
         exif_bytes = piexif.dump(exif_dict)
         piexif.insert(exif_bytes, path)
     except (ValueError, InvalidImageDataError):
-        logger = setup_logger()
-        logger.debug("Error setting EXIF data for %s", path)
+        logger.tqdm_write(f"Error setting EXIF data for {path}", logging.DEBUG)
         return