+++ tests/test_download_photos.py
@@ -82,7 +82,7 @@ class DownloadPhotoTestCase(TestCase):
                 self._caplog.text,
             )
             self.assertIn(
-                f"INFO     Downloading {os.path.join(base_dir, os.path.normpath('2018/07/31/IMG_7409.JPG'))}",
+                f"DEBUG    Downloading {os.path.join(base_dir, os.path.normpath('2018/07/31/IMG_7409.JPG'))}",
                 self._caplog.text,
             )
             self.assertNotIn(
@@ -98,11 +98,11 @@ class DownloadPhotoTestCase(TestCase):
                 self._caplog.text,
             )
             self.assertIn(
-                "INFO     Skipping IMG_7405.MOV, only downloading photos.",
+                "DEBUG    Skipping IMG_7405.MOV, only downloading photos.",
                 self._caplog.text,
             )
             self.assertIn(
-                "INFO     Skipping IMG_7404.MOV, only downloading photos.",
+                "DEBUG    Skipping IMG_7404.MOV, only downloading photos.",
                 self._caplog.text,
             )
             self.assertIn(
@@ -202,7 +202,7 @@ class DownloadPhotoTestCase(TestCase):
                         self._caplog.text,
                     )
                     self.assertIn(
-                        f"INFO     Downloading {os.path.join(base_dir, os.path.normpath('2018/07/31/IMG_7409.JPG'))}",
+                        f"DEBUG    Downloading {os.path.join(base_dir, os.path.normpath('2018/07/31/IMG_7409.JPG'))}",
                         self._caplog.text,
                     )
                     # 2018:07:31 07:22:24 utc
@@ -271,7 +271,7 @@ class DownloadPhotoTestCase(TestCase):
                     self._caplog.text,
                 )
                 self.assertIn(
-                    f"INFO     Downloading {os.path.join(base_dir, os.path.normpath('2018/07/31/IMG_7409.JPG'))}",
+                    f"DEBUG    Downloading {os.path.join(base_dir, os.path.normpath('2018/07/31/IMG_7409.JPG'))}",
                     self._caplog.text,
                 )
                 self.assertIn(
@@ -435,7 +435,7 @@ class DownloadPhotoTestCase(TestCase):
                     expected_calls = list(
                         map(
                             lambda f: call(
-                                ANY, ANY, os.path.join(
+                                ANY, False, ANY, ANY, os.path.join(
                                     base_dir, os.path.normpath(f[0])),
                                 "mediumVideo" if (
                                     f[1] == 'photo' and f[0].endswith('.MOV')
@@ -588,7 +588,7 @@ class DownloadPhotoTestCase(TestCase):
                         )
 
                         self.assertIn(
-                            "INFO     Could not download IMG_7409.JPG! Please try again later.",
+                            "ERROR    Could not download IMG_7409.JPG! Please try again later.",
                             self._caplog.text,
                         )
 
@@ -730,7 +730,7 @@ class DownloadPhotoTestCase(TestCase):
                         )
 
                         self.assertIn(
-                            "INFO     Could not download IMG_7409.JPG! Please try again later.",
+                            "ERROR    Could not download IMG_7409.JPG! Please try again later.",
                             self._caplog.text,
                         )
                         assert result.exit_code == 0
@@ -910,13 +910,15 @@ class DownloadPhotoTestCase(TestCase):
                             self._caplog.text,
                         )
                         self.assertIn(
-                            f"INFO     Downloading {os.path.join(base_dir, os.path.normpath('2018/07/31/IMG_7409.JPG'))}",
+                            f"DEBUG    Downloading {os.path.join(base_dir, os.path.normpath('2018/07/31/IMG_7409.JPG'))}",
                             self._caplog.text,
                         )
                         self.assertIn(
                             "INFO     All photos have been downloaded!", self._caplog.text
                         )
                         dp_patched.assert_called_once_with(
+                            ANY,
+                            False,
                             ANY,
                             ANY,
                             f"{os.path.join(base_dir, os.path.normpath('2018/07/31/IMG_7409.JPG'))}",
@@ -1043,7 +1045,7 @@ class DownloadPhotoTestCase(TestCase):
                     self._caplog.text,
                 )
                 self.assertIn(
-                    f"INFO     Downloading {os.path.join(base_dir, os.path.normpath('2018/01/01/IMG_7409.JPG'))}",
+                    f"DEBUG    Downloading {os.path.join(base_dir, os.path.normpath('2018/01/01/IMG_7409.JPG'))}",
                     self._caplog.text,
                 )
                 self.assertIn(
@@ -1117,7 +1119,7 @@ class DownloadPhotoTestCase(TestCase):
                     self._caplog.text,
                 )
                 self.assertIn(
-                    f"INFO     Downloading {os.path.join(base_dir, os.path.normpath('5/01/01/IMG_7409.JPG'))}",
+                    f"DEBUG    Downloading {os.path.join(base_dir, os.path.normpath('5/01/01/IMG_7409.JPG'))}",
                     self._caplog.text,
                 )
                 self.assertIn(
@@ -1176,7 +1178,7 @@ class DownloadPhotoTestCase(TestCase):
                         self._caplog.text,
                     )
                     self.assertIn(
-                        "INFO     Skipping IMG_7409.JPG, only downloading photos and videos. (Item type was: unknown)",
+                        "DEBUG    Skipping IMG_7409.JPG, only downloading photos and videos. (Item type was: unknown)",
                         self._caplog.text,
                     )
                     self.assertIn(
@@ -1253,15 +1255,15 @@ class DownloadPhotoTestCase(TestCase):
                     self._caplog.text,
                 )
                 self.assertIn(
-                    f"INFO     Downloading {os.path.join(base_dir, os.path.normpath('2018/07/31/IMG_7409-1884695.JPG'))}",
+                    f"DEBUG    Downloading {os.path.join(base_dir, os.path.normpath('2018/07/31/IMG_7409-1884695.JPG'))}",
                     self._caplog.text,
                 )
                 self.assertIn(
-                    f"INFO     {os.path.join(base_dir, os.path.normpath('2018/07/31/IMG_7409-3294075.MOV'))} deduplicated.",
+                    f"DEBUG    {os.path.join(base_dir, os.path.normpath('2018/07/31/IMG_7409-3294075.MOV'))} deduplicated.",
                     self._caplog.text,
                 )
                 self.assertIn(
-                    f"INFO     Downloading {os.path.join(base_dir, os.path.normpath('2018/07/31/IMG_7409-3294075.MOV'))}",
+                    f"DEBUG    Downloading {os.path.join(base_dir, os.path.normpath('2018/07/31/IMG_7409-3294075.MOV'))}",
                     self._caplog.text,
                 )
                 self.assertIn(
@@ -1273,10 +1275,10 @@ class DownloadPhotoTestCase(TestCase):
                     self._caplog.text,
                 )
                 self.assertIn(
-                    "INFO     Skipping IMG_7405.MOV, only downloading photos.", self._caplog.text
+                    "DEBUG    Skipping IMG_7405.MOV, only downloading photos.", self._caplog.text
                 )
                 self.assertIn(
-                    "INFO     Skipping IMG_7404.MOV, only downloading photos.", self._caplog.text
+                    "DEBUG    Skipping IMG_7404.MOV, only downloading photos.", self._caplog.text
                 )
                 self.assertIn(
                     "INFO     All photos have been downloaded!", self._caplog.text
@@ -1352,7 +1354,7 @@ class DownloadPhotoTestCase(TestCase):
                         self._caplog.text,
                     )
                     self.assertIn(
-                        f"INFO     Downloading {os.path.join(base_dir, os.path.normpath('2018/07/31/IMG_7409.JPG'))}",
+                        f"DEBUG    Downloading {os.path.join(base_dir, os.path.normpath('2018/07/31/IMG_7409.JPG'))}",
                         self._caplog.text,
                     )
                     # 2018:07:31 07:22:24 utc
@@ -1421,7 +1423,7 @@ class DownloadPhotoTestCase(TestCase):
                 self._caplog.text,
             )
             self.assertIn(
-                f"INFO     Downloading {os.path.join(base_dir, os.path.normpath('2018/07/31/IMG_7409.JPG'))}",
+                f"DEBUG    Downloading {os.path.join(base_dir, os.path.normpath('2018/07/31/IMG_7409.JPG'))}",
                 self._caplog.text,
             )
             self.assertNotIn(
@@ -1502,11 +1504,11 @@ class DownloadPhotoTestCase(TestCase):
                         self._caplog.text,
                     )
                     self.assertIn(
-                        f"INFO     Downloading {os.path.join(base_dir, os.path.normpath('2018/07/31/IMG_7409.JPG'))}",
+                        f"DEBUG    Downloading {os.path.join(base_dir, os.path.normpath('2018/07/31/IMG_7409.JPG'))}",
                         self._caplog.text,
                     )
                     self.assertIn(
-                        "INFO     Deleting IMG_7409.JPG", self._caplog.text
+                        "INFO     Deleted IMG_7409.JPG in iCloud", self._caplog.text
                     )
                     self.assertIn(
                         "INFO     All photos have been downloaded!", self._caplog.text
@@ -1560,11 +1562,11 @@ class DownloadPhotoTestCase(TestCase):
                 self._caplog.text,
             )
             self.assertIn(
-                f"INFO     Downloading {os.path.join(base_dir, os.path.normpath('2018/07/31/IMG_7409.JPG'))}",
+                f"DEBUG    Downloading {os.path.join(base_dir, os.path.normpath('2018/07/31/IMG_7409.JPG'))}",
                 self._caplog.text,
             )
             self.assertNotIn(
-                "INFO     Deleting IMG_7409.JPG", self._caplog.text
+                "INFO     Deleted IMG_7409.JPG in iCloud", self._caplog.text
             )
             self.assertIn(
                 "INFO     All photos have been downloaded!", self._caplog.text
@@ -1628,7 +1630,7 @@ class DownloadPhotoTestCase(TestCase):
                 self._caplog.text,
             )
             self.assertIn(
-                f"INFO     Downloading {os.path.join(base_dir, os.path.normpath('2018/07/31/IMG_7409.JPG'))}",
+                f"DEBUG    Downloading {os.path.join(base_dir, os.path.normpath('2018/07/31/IMG_7409.JPG'))}",
                 self._caplog.text,
             )
             self.assertNotIn(
@@ -1644,11 +1646,11 @@ class DownloadPhotoTestCase(TestCase):
                 self._caplog.text,
             )
             self.assertIn(
-                "INFO     Skipping IMG_7405.MOV, only downloading photos.",
+                "DEBUG    Skipping IMG_7405.MOV, only downloading photos.",
                 self._caplog.text,
             )
             self.assertIn(
-                "INFO     Skipping IMG_7404.MOV, only downloading photos.",
+                "DEBUG    Skipping IMG_7404.MOV, only downloading photos.",
                 self._caplog.text,
             )
             self.assertIn(
@@ -1856,7 +1858,7 @@ class DownloadPhotoTestCase(TestCase):
                     # )
 
                     self.assertIn(
-                        "INFO     Could not download IMG_7409.JPG! Please try again later.",
+                        "ERROR    Could not download IMG_7409.JPG! Please try again later.",
                         self._caplog.text,
                     )
 
@@ -1928,3 +1930,196 @@ class DownloadPhotoTestCase(TestCase):
             base_dir, "**/*.*"), recursive=True)
 
         assert sum(1 for _ in files_in_result) == 0
+
+    def test_handle_io_error_mkdir(self):
+        base_dir = os.path.join(self.fixtures_path, inspect.stack()[0][3])
+        recreate_path(base_dir)
+
+        with vcr.use_cassette(os.path.join(self.vcr_path, "listing_photos.yml")):
+            with mock.patch("os.makedirs", create=True) as m:
+                # Raise IOError when we try to write to the destination file
+                m.side_effect = IOError
+
+                runner = CliRunner(env={
+                    "CLIENT_ID": "DE309E26-942E-11E8-92F5-14109FE0B321"
+                })
+                result = runner.invoke(
+                    main,
+                    [
+                        "--username",
+                        "jdoe@gmail.com",
+                        "--password",
+                        "password1",
+                        "--recent",
+                        "1",
+                        "--skip-videos",
+                        "--skip-live-photos",
+                        "--no-progress-bar",
+                        "--threads-num",
+                        1,
+                        "-d",
+                        base_dir,
+                    ],
+                )
+                print_result_exception(result)
+
+                self.assertIn(
+                    "DEBUG    Looking up all photos from album All Photos...", self._caplog.text)
+                self.assertIn(
+                    f"INFO     Downloading the first original photo to {base_dir} ...",
+                    self._caplog.text,
+                )
+                self.assertIn(
+                    f"ERROR    Could not create folder {base_dir}",
+                    self._caplog.text,
+                )
+                self.assertEqual(result.exit_code, 0, "Exit code")
+
+        files_in_result = glob.glob(os.path.join(
+            base_dir, "**/*.*"), recursive=True)
+
+        self.assertEqual(sum(1 for _ in files_in_result), 0, "Files at the end")
+
+    def test_dry_run(self):
+        base_dir = os.path.join(self.fixtures_path, inspect.stack()[0][3])
+        recreate_path(base_dir)
+
+        files_to_download = [
+            '2018/07/31/IMG_7409.JPG',
+            # "2018/07/30/IMG_7408.JPG",
+            # "2018/07/30/IMG_7407.JPG",
+        ]
+
+        with vcr.use_cassette(os.path.join(self.vcr_path, "listing_photos.yml")):
+            # Pass fixed client ID via environment variable
+            runner = CliRunner(env={
+                "CLIENT_ID": "DE309E26-942E-11E8-92F5-14109FE0B321"
+            })
+            result = runner.invoke(
+                main,
+                [
+                    "--username",
+                    "jdoe@gmail.com",
+                    "--password",
+                    "password1",
+                    "--recent",
+                    "1",
+                    "--skip-videos",
+                    "--skip-live-photos",
+                    "--set-exif-datetime",
+                    "--no-progress-bar",
+                    "--dry-run",
+                    "--threads-num",
+                    1,
+                    "-d",
+                    base_dir,
+                ],
+            )
+            print_result_exception(result)
+
+            self.assertIn(
+                "DEBUG    Looking up all photos from album All Photos...", self._caplog.text)
+            # self.assertIn(
+            #     f"INFO     Downloading 2 original photos to {base_dir} ...",
+            #     self._caplog.text,
+            # )
+            for f in files_to_download:
+                self.assertIn(
+                    f"DEBUG    Downloading {os.path.join(base_dir, os.path.normpath(f))}",
+                    self._caplog.text,
+                )
+            self.assertNotIn(
+                "IMG_7409.MOV",
+                self._caplog.text,
+            )
+            self.assertNotIn(
+                "ERROR",
+                self._caplog.text,
+            )
+            # self.assertIn(
+            #     "DEBUG    Skipping IMG_7405.MOV, only downloading photos.",
+            #     self._caplog.text,
+            # )
+            # self.assertIn(
+            #     "DEBUG    Skipping IMG_7404.MOV, only downloading photos.",
+            #     self._caplog.text,
+            # )
+            self.assertIn(
+                "INFO     All photos have been downloaded!", self._caplog.text
+            )
+
+            assert result.exit_code == 0
+
+        files_in_result = glob.glob(os.path.join(
+            base_dir, "**/*.*"), recursive=True)
+
+        self.assertEqual(sum(1 for _ in files_in_result), 0, "Files in the result")
+
+    def test_download_after_delete_dry_run(self):
+        base_dir = os.path.join(self.fixtures_path, inspect.stack()[0][3])
+        recreate_path(base_dir)
+
+        def raise_response_error(a0_, a1_, a2_):
+            raise Exception("Unexpected call to delete_photo")
+
+        with mock.patch.object(piexif, "insert") as piexif_patched:
+            piexif_patched.side_effect = InvalidImageDataError
+            with mock.patch(
+                "icloudpd.exif_datetime.get_photo_exif"
+            ) as get_exif_patched:
+                get_exif_patched.return_value = False
+                with mock.patch(
+                    "icloudpd.base.delete_photo"
+                ) as df_patched:
+                    df_patched.side_effect = raise_response_error
+
+                    with vcr.use_cassette(os.path.join(self.vcr_path, "listing_photos.yml")) as cass:
+                        # Pass fixed client ID via environment variable
+                        runner = CliRunner(env={
+                            "CLIENT_ID": "DE309E26-942E-11E8-92F5-14109FE0B321"
+                        })
+                        result = runner.invoke(
+                            main,
+                            [
+                                "--username",
+                                "jdoe@gmail.com",
+                                "--password",
+                                "password1",
+                                "--recent",
+                                "1",
+                                "--skip-videos",
+                                "--skip-live-photos",
+                                "--no-progress-bar",
+                                "--dry-run",
+                                "--threads-num",
+                                1,
+                                "--delete-after-download",
+                                "-d",
+                                base_dir,
+                            ],
+                        )
+                        print_result_exception(result)
+
+                        self.assertIn(
+                            "DEBUG    Looking up all photos from album All Photos...", self._caplog.text)
+                        self.assertIn(
+                            f"INFO     Downloading the first original photo to {base_dir} ...",
+                            self._caplog.text,
+                        )
+                        self.assertIn(
+                            f"DEBUG    Downloading {os.path.join(base_dir, os.path.normpath('2018/07/31/IMG_7409.JPG'))}",
+                            self._caplog.text,
+                        )
+                        self.assertIn(
+                            "INFO     [DRY RUN] Would delete IMG_7409.JPG in iCloud", self._caplog.text
+                        )
+                        self.assertIn(
+                            "INFO     All photos have been downloaded!", self._caplog.text
+                        )
+                        self.assertEqual(cass.all_played, False, "All mocks played")
+                        self.assertEqual(result.exit_code, 0, "Exit code")
+
+        files_in_result = glob.glob(os.path.join(
+            base_dir, "**/*.*"), recursive=True)
+
+        self.assertEqual( sum(1 for _ in files_in_result), 0, "Files in the result")